python-ml/requirements.txt
nats-py>=2.6.0
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0
hmmlearn>=0.3.0
lightgbm>=4.0.0
arch>=6.0.0
torch>=2.1.0
dowhy>=0.11
econml>=0.15
statsmodels>=0.14.0
yfinance>=0.2.40
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
structlog>=24.1.0
